Use hyperlinks when mentioning bips
When mentioning a bip we use a hyperlink to the bip file to assist
readers. There are two mentions of bips in BIP-0009 that do not use
links, probably because links where provided above in the text, however,
its a bit easier on the reader if we provide links again to save
scrolling through the document.

Use hyperlinks when mentioning BIP 34 and 141 in isolation.